requirements.txt
# Core Libraries
requests==2.32.3
pandas==2.2.2
numpy==1.26.0
tqdm==4.66.5

# Web Scraping
beautifulsoup4==4.12.3

# Visualization
matplotlib==3.6.0
seaborn==0.12.0
WordCloud==1.9.3
bertopic==0.16.3

# Development Tools
python-dotenv==1.0.1
ipykernel==6.29.5
pylint==3.3.1
autopep8==2.3.1